==> 01_analysis_pre_matching/02_mentoring_eligibility_check_25.Rmd <==
---
title: "Researcher Mentoring Program for Early Career Researchers (2025)"
# subtitle: "Eligibility Check"
# author: "Jen Beaudry"
date: "January 2025"
output: pdf_document
editor_options: 
  chunk_output_type: console
header-includes:
  - \usepackage{titling}
    \setlength{\droptitle}{-2cm} # removes some of the white space above title
  - \pretitle{\begin{center}
    \includegraphics[width=1.5in, height=1.5in]{flinders-university-horizontal-master.png}\LARGE\\} # adds logo
  - \posttitle{\end{center}}
  - \usepackage{fancyhdr}
  - \fancypagestyle{plain}{\pagestyle{fancy}} # sets the footer on page 1
  - \pagestyle{fancy} # sets the footer on page 2 and beyond
  - \fancyhf{} # clears existing header & footer
  - \renewcommand{\headrulewidth}{0pt} # removes the line from the header
  - \fancyfoot[L]{Jen Beaudry} # my name in footer
  - \fancyfoot[R]{\thepage} # page number in foooter
---

```{r setup, include=FALSE}

knitr::opts_chunk$set(echo = FALSE, include = TRUE)

options(knitr.duplicate.label = "allow") # use this when you use the child files
```

```{r library, include=FALSE}

library(here)
library(tidyverse)
library(kableExtra)
library(readxl)

```

```{r run preprocessing, child=(here::here("01_analysis_pre_matching", "01_mentoring_eoi_preprocessing_25.Rmd")), include=FALSE}


```

```{r load data, include=FALSE}

# mentor data

df_tor <- read_csv(here::here("00_data", "processed_data", "eoi_mentor_preprocessed.csv"), lazy = FALSE)

# mentee data

df_tee <- read_csv(here::here("00_data", "processed_data", "eoi_mentee_preprocessed.csv"), lazy = FALSE)

# load metadata for mentor

meta_tor <- read_csv(here::here("00_data", "raw_data", "metadata_eoi_mentor.csv"), lazy = FALSE) %>%
  filter(old_variable != "NA", old_variable != "exclude") # remove the instruction variables

# load metadata for mentee

meta_tee <- read_csv(here::here("00_data", "raw_data", "metadata_eoi_mentee.csv"), lazy = FALSE) %>%
  filter(old_variable != "NA", old_variable != "exclude") # remove the instruction variables

```

```{r functions, include=FALSE}

styler_pdf <- function(dat) {
  dat %>%
     kable_styling(full_width = F, 
                latex_options = c("striped", "HOLD_position"))
}

source(here("..", "functions", "missing_data.R"))

```

```{r code factors, include=FALSE}

df_tor$college <- as.factor(df_tor$college)
df_tee$college <- as.factor(df_tee$college)

df_tor$id <- as.factor(df_tor$id)
df_tee$id <- as.factor(df_tee$id)


```

```{r check for duplicates, include = FALSE}

# [note: the reason this is here rather than preprocessing is so I can use the 
  # data for reporting; if this chunk is updated, also update it in 00_ file]

# get our starting numbers

original_n_tee <- nrow(df_tee)
original_n_tor <- nrow(df_tor)

# we then need to create values for 'dup_tee' & 'dup_tor' that we can use in the
  # report, if there are any duplicates. Don't need if there are no duplicates.

# MENTORS

# check for duplicates in the data
  # integer(0) means there are no duplicates
  # any other number reflects a duplicate. Two or more numbers indicates two or
  # more duplicates. 

which(duplicated(df_tor$surname))

# if no duplicate, add a duplicate variable so it matches the mentee data

# two possible duplicate flaggeds; two real duplicates 
  # delete mentor_11; mentor_32 was the complete application
  # delete mentor_18; mentor_49 was the complete application


# after reviewing the raw data, indicate duplicates in the df
# if we have more than one, add them in like this:  ((id %in% "mentee_25") | (id %in% "mentee_23")


df_tor <- df_tor %>%
  mutate (duplicate = ifelse ((id %in% "mentor_11") |
                               id %in% "mentor_18",
                              "yes",
                              "no"
  )) 

# create object indicating how many mentors submitted duplicate EOIs(needed for report)

dup_tor <- df_tor %>% 
  filter(duplicate == "yes") %>% 
  nrow()

# remove duplicate from working df

df_tor <- df_tor %>% 
  filter(duplicate == "no")

# double check by looking at first name duplicates before moving forward
which(duplicated(df_tor$first_name))

# four people share the same first names (two pairs)


# MENTEES

which(duplicated(df_tee$surname))

# five possible duplicates flagged; five real duplicates 
  # delete mentee_23; mentee_24 was the complete application
  # delete mentee_11; mentee_41 was the complete application
  # delete mentee_19; mentee_46 was the complete application
  # delete mentee_27; mentee_52 was the complete application
  # delete mentee_33; mentee_53 was the complete application

# after reviewing the raw data, indicate duplicates in the df


df_tee <- df_tee %>%
  mutate (duplicate = ifelse ((id %in% "mentee_11") | 
                                (id %in% "mentee_19") | 
                                (id %in% "mentee_23") |
                                (id %in% "mentee_27") | 
                                (id %in% "mentee_33"),
                              "yes",
                              "no"
  )) 

# create object indicating how many mentees submitted duplicate EOIs(needed for report)

dup_tee <- df_tee %>% 
  filter(duplicate == "yes") %>% 
  nrow()

# remove duplicate from working df

df_tee <- df_tee %>% 
  filter(duplicate == "no")

# double check by looking at first name duplicates before moving forward. No shared names. 
which(duplicated(df_tee$first_name))


```


```{r withdrawals and errors, include = FALSE}

## MENTEES

# 2 withdrawals this year [1 after matching]

withdraw_n_tee_pre <- df_tee %>%
    filter(id %in% c("mentee_38")) %>% 
    nrow()

withdraw_n_tee_post <- df_tee %>%
    filter(id %in% c("mentee_4")) %>% 
    nrow()
  
# remove the mentees from the df

df_tee <- df_tee %>%
  filter(!id %in% c("mentee_38", "mentee_4"))


## MENTORS

# 1 withdrawal this year [we did not force withdrawal of the mentor whose mentee withdrew post matching]

withdraw_n_tor_pre <- df_tor %>%
    filter(id %in% c("mentor_16")) %>%
    nrow()

# remove the mentor from the df

df_tor <- df_tor %>%
  filter(!id %in% c("mentor_16"))

```


```{r code eligibility criteria mentors, include = FALSE}

# Need to explain who we've lost along the way 
  # [if we end up with multiple mentors who aren't eligible and
  # overlap across criteria, then work through the steps like I do with mentees.
  # For now this is fine for this cohort of mentors.]

## ACADEMIC LEVEL ##

# filter all of the cases that have info for "other_level"
# use this dataframe to look at these cases and figure out what needs to be done first, 
  # then work through the steps below

other_level <- df_tor %>%
  filter(level %in% "Other (please specify)") 

df_tor$level <- as.factor(df_tor$level)

df_tor <- df_tor %>%
  mutate(level_eligibility = case_when(id %in% "mentor_9" ~ "Not eligible", # level B
                                       id %in% "mentor_41" ~ "Not eligible", # level B
                                 TRUE ~ as.character("Eligible"))) %>% 
  mutate(level_eligibility = as.factor(level_eligibility))

# not needed in 2025

# after looking thru the data, three cases need to be recoded because two were 
  # promoted as of Jan 2023 and one is a level B [removed from data earlier]. 
  # All other cases will copy over the original level value

# df_tor <- df_tor %>%
#   mutate(level_recode = case_when(id %in% "mentor_9" ~ "Level D", 
#                                   id %in% "mentor_10" ~ "Level E",
#                                   id %in% "mentor_28" ~ "Level B", # removed from data
#                                   id %in% "mentor_12" ~ "Level E",
#                                  TRUE ~ as.character(level))) %>% 
#   relocate(level_recode, .after = level_other)



# create object for report of how many are not eligible due to academic level. none.

level_no_n_tor <- df_tor %>%
    filter(level_eligibility == "Not eligible") %>%
    nrow()

# if no levels need to be recoded, add the recode variable anyway to ensure that 
  # the mentor and mentee dataframes match

df_tor <- df_tor %>%
  mutate(level_recode = level)

## PART-TIME EMPLOYMENT ##

# now look at whether there is anyone with a different employment type or 
  # working less than 0.4

df_tor %>%
  filter(employment_type %in% "Other (please specify)") 

# no one with any other type of employment, so nothing to do here

# let's rename the part-time value with simpler text (remove the 'what is your FTE?)

df_tor <- df_tor %>%
  mutate(
    employment_type = replace(
      employment_type,
      str_detect(employment_type, "Part-time"),
      "Part-time"
    )
  ) 

# confirm that no one has less than 0.4

df_tor %>% 
  mutate(as.numeric(parttime_FTE)) %>% 
  filter(parttime_FTE < 0.4)

# no mentors with less than 0.4 FTE, so nothing more to do. Indicate eligibility here.

# add an employment eligibility variable to df

df_tor <- df_tor %>%
  mutate(employment_eligibility = as.character("Eligible")) %>%
  mutate(employment_eligibility = as.factor(employment_eligibility))

# create object for report of how many are not eligible due to employment type

employ_no_n_tor <- df_tor %>% 
    filter(employment_eligibility == "Not eligible") %>% 
    nrow()



## CONTRACT END DATE ##

# Manually review their end date....these are mentors with fixed-term contracts

df_tor_end <- df_tor %>% 
   filter(str_detect(employment_length, "Fixed-term"))

# I did this manually after screening by contract end date & level
# if necessary, contact those who did not provide contract end date details [not necessary in 2025]


# add a contract eligibility variable to df


df_tor <- df_tor %>%
  mutate(contract_eligibility = as.character("Eligible")) %>%
  mutate(contract_eligibility = as.factor(contract_eligibility))

# create object for report of how many are not eligible due to contract

contract_no_n_tor <- df_tor %>% 
    filter(!contract_eligibility == "Eligible") %>% #takes into account the question about eligibility
    nrow()


## CODE OVERALL ELIGIBILITY ###

df_tor <- df_tor %>%
  mutate(eligibility = case_when(employment_eligibility == "Not eligible" ~ "Not eligible", 
                                level_eligibility == "Not eligible" ~ "Not eligible", 
                                contract_eligibility == "Not eligible" ~ "Not eligible", 
                                #grepl("Question", contract_eligibility) ~ "Information needed", {not needed}
                                  TRUE ~ as.character("Eligible"))) %>% 
  mutate(eligibility = as.factor(eligibility))

# keep everyone in the dataframe for now so we have the full dataset. Will remove
# ineligible folks later

rm(other_level, df_tor_end)

```

```{r code eligibility criteria mentees, include = FALSE}

# Need to explain who we've lost along the way (see internal_notes.txt)

## PREVIOUS PARTICIPATION IN PROGRAM ##

# check to see if this year's applicants previously participated in the program. 
  # use code below to check the data with all prior mentees against
  # this year's applicants. Find the updated spreadsheet in our general T&D files.

# load the excel sheet with all prior participants and select only the mentee column
df_prior <- read_xlsx(
  here::here("00_data",
             "raw_data",
             "all_participants_since_2012.xlsx"),
  sheet = 'Names',
#  n_max = 129,  #
#  .name_repair = "unique",
  trim_ws = TRUE) %>% 
  select (Mentee, Year) 

# select only the mentee's names for this year's applicants
df_tee_name <- df_tee %>% 
  select (c(id, name, college)) %>% 
  rename ("Mentee"= "name")

df_compare <- inner_join(df_prior, df_tee_name, by = "Mentee")


# one mentee was previously selected for the program so is ineligible to 
  # participate this year

df_tee <- df_tee %>%
  mutate(previous_mentee_eligibility = case_when(id %in% "mentee_42" ~ "Not eligible", # participated in 2022
                                           TRUE ~ as.character("Eligible"))) %>% 
  mutate(previous_mentee_eligibility = as.factor(previous_mentee_eligibility))

previous_no_n_tee <- df_tee %>% 
    filter(previous_mentee_eligibility == "Not eligible") %>%
    nrow()

rm(df_compare, df_tee_name, df_prior)

## EMPLOYMENT TYPE ##

# look at their responses to other employment types AND for those in other academic roles AND
  # for those who said that their position was a PhD student

employ_a <- df_tee %>%
  filter(employment_type %in% "Other (please specify)") 

employ_b <- df_tee %>%
  filter(academic_role %in% "Other (please specify)") 

employ_c <- df_tee %>% 
  filter(str_detect(position, "student"))

# mark them as ineligible as required 
df_tee <- df_tee %>%
  mutate(employment_eligibility = case_when(id %in% "mentee_5" ~ "Not eligible", # PhD student & Research Assistant
                                            id %in% "mentee_44" ~ "Not eligible", # PhD student
                                            id %in% "mentee_18" ~ "Not eligible", # PhD student
                                            id %in% "mentee_55" ~ "Not eligible", # PhD student
                                            TRUE ~ as.character("Eligible"))) %>% 
  mutate(employment_eligibility = as.factor(employment_eligibility))

# create object for report of how many are not eligible due to employment type

employ_no_n_tee <- df_tee %>% 
    filter(employment_eligibility == "Not eligible") %>%
    nrow()

rm (employ_a, employ_b, employ_c)

# PART-TIME EMPLOYMENT

# let's rename the part-time value with simpler text

df_tee <- df_tee %>%
  mutate(
    employment_type = replace(
      employment_type,
      str_detect(employment_type, "Part-time"),
      "Part-time"
    )
  ) 

# confirm that no one has less than 0.4 [check the values to see if anything needs 
  # to be recoded first]

df_tee %>% 
  mutate(as.numeric(parttime_FTE)) %>% 
  filter(parttime_FTE < 0.4)

# no mentees with less than 0.4 FTE, so nothing more to do


## ACADEMIC LEVEL FOR MENTEES ##

# look at the cases that have info for "other_level" and then recode the levels. 
  # if their levels are fine, copy over the level value. 

level <- df_tee %>%
  filter(level %in% "Other (please specify)")

# see if any ineligible levels exist. none in this dataset.
df_tee$level <- as.factor(df_tee$level)
df_tee %>% 
  filter(!level %in% c("Level A", "Level B", "Level C"))
# no one other than those with other levels because 
  # technically Level Ds can participate according to the guidelines!

# this list also includes others who were revealed to have casual appts based 
  # on other info in their EOI (e.g., mentee_43)

df_tee <- df_tee %>%
  mutate(level_recode = case_when(id %in% "mentee_5" ~ "Research Assistant",
                                  id %in% "mentee_18" ~ "PhD student", 
                                  id %in% "mentee_44" ~ "PhD student",
                                  id %in% "mentee_55" ~ "PhD student",
                                 TRUE ~ as.character((level)))) %>% 
  relocate(level_recode, .after = level_other)

# now mark eligibility according to level

df_tee <- df_tee %>%
  mutate(level_eligibility = case_when(id %in% "mentee_5" ~ "Not eligible", 
                                  id %in% "mentee_18" ~ "Not eligible",
                                  id %in% "mentee_44" ~ "Not eligible",
                                  id %in% "mentee_55" ~ "Not eligible",
                                 TRUE ~ as.character("Eligible"))) %>% 
  mutate(level_eligibility = as.factor(level_eligibility))


# create object for report of how many are not eligible due to previous participation or level, considering 
  # only folks who are eligible according to employment type


level_no_n_tee <- df_tee %>% 
    filter(previous_mentee_eligibility == "Eligible") %>% 
    filter(employment_eligibility == "Eligible") %>% 
    filter(!level_eligibility %in% "Eligible") %>% 
    nrow()

rm (level)

### PhD CONFERRAL DATE ###

# manually review their conferral date [can sort by year]
df_phd <- df_tee %>% 
  select(c(id, name, college, conferral_month, conferral_year, interruptions, employment_eligibility, level_eligibility))

# now mark eligibility according to conferral date
  # [breadcrumb: resolve awaiting information eligibility issues]
  
df_tee <- df_tee %>%
  mutate(conferral_eligibility = case_when(id %in% "mentee_9" ~ "Not eligible", # does not yet hold a phd
                                  id %in% "mentee_20" ~ "Not eligible", # email re: career interruptions
                                 TRUE ~ as.character("Eligible"))) %>% 
  mutate(conferral_eligibility = as.factor(conferral_eligibility))


# create object for report of how many are not eligible due to conferral date [or 
  # ones that we still need more info from], considering 
  # only folks who are eligible according to employment type and academic level

phd_no_n_tee <- df_tee %>% 
    filter(previous_mentee_eligibility == "Eligible") %>% 
    filter(employment_eligibility == "Eligible") %>% 
    filter(level_eligibility == "Eligible") %>% 
    filter(conferral_eligibility == "Not eligible" | conferral_eligibility == "Information needed") %>% 
    nrow()

remove(df_phd)

### CONTRACT END DATE ###

# Manually review their end date....these are mentees with fixed-term contracts
  # code is valid, but we don't need to run it each time

# let's rename the fixed term value with simpler text

df_tee <- df_tee %>%
  mutate(
    employment_length = replace(
      employment_length,
      str_detect(employment_length, "Fixed-term"),
      "Fixed-term"
    )
  ) 

# manually review their contract end date 

df_tee_end <- df_tee %>% 
  filter(str_detect(employment_length, "Fixed-term")) %>% 
  select(c(id, name, college, employment_length, contract_end_date, level_recode, employment_eligibility, level_eligibility, conferral_eligibility))

# I did this manually after screening by contract end date & level
# then I reviewed their EOIs and contacted those who did not provide contract end date details
# this also captures those above who were students or casuals, but I'm not including those here

df_tee <- df_tee %>%
  mutate(contract_eligibility = case_when(id %in% "mentee_5" ~ "Not eligible", # email
                                 id %in% "mentee_9" ~ "Not eligible", # email
                                 id %in% "mentee_15" ~ "Eligible", # contract extended [via email]
                                 id %in% "mentee_31" ~ "Eligible", # contract until end of year [via email]
                                 id %in% "mentee_49" ~ "Eligible", # contract extended [via email]
                                 id %in% "mentee_52" ~ "Eligible", # contract extended [via email]
                                 id %in% "mentee_54" ~ "Not eligible", # incomplete application too
                                  TRUE ~ as.character("Eligible"))) %>% 
  mutate(contract_eligibility = as.factor(contract_eligibility))

# create object for report of how many are not eligible due to contract end date 
  # considering only folks who are eligible according to employment type, academic level, 
  # and conferral date

contract_no_n_tee <- df_tee %>% 
    filter(previous_mentee_eligibility == "Eligible") %>% 
    filter(employment_eligibility == "Eligible") %>% 
    filter(level_eligibility == "Eligible") %>% 
    filter(conferral_eligibility == "Eligible") %>% 
    filter(contract_eligibility  == "Not eligible" | contract_eligibility == "Information needed") %>% 
    nrow()


## CREATE OVERALL ELIGIBILITY VARIABLE ## 
# keep the 'information needed' text because that will be helpful when I am in the middle of deciding eligibility

df_tee <- df_tee %>%
  mutate(
    eligibility = case_when(
      previous_mentee_eligibility == "Not eligible" ~ "Not eligible", 
                                employment_eligibility == "Not eligible" ~ "Not eligible", 
                                level_eligibility == "Not eligible" ~ "Not eligible", 
                                conferral_eligibility == "Not eligible" ~ "Not eligible", 
                               contract_eligibility == "Not eligible" ~ "Not eligible", 
                                employment_eligibility  == "Information needed" ~ "Information needed", 
                               level_eligibility  == "Information needed" ~ "Information needed", 
                               conferral_eligibility  == "Information needed" ~ "Information needed", 
                             contract_eligibility == "Information needed" ~ "Information needed", 
                                  TRUE ~ as.character("Eligible"))) %>% 
  mutate(eligibility = as.factor(eligibility))


# keep everyone in the dataframe for now so we have the full dataset. Will remove
# ineligible folks later

rm (df_tee_end)

```

```{r save processed data, include = FALSE}

# when done assessing for eligibility, write the data to new files
# these files include eligible and ineligible folks, with eligibility status indicated in the file
# row.names gets rid of the first column from the dataframe.

write.csv(df_tor, here::here("00_data", "processed_data", "eoi_mentor_processed.csv"), row.names = FALSE)

write.csv(df_tee, here::here("00_data", "processed_data", "eoi_mentee_processed.csv"), row.names = FALSE)

# save a subset with details about ineligible folks and those with outstanding questions
df_tee_inel <- df_tee %>% 
  filter(!eligibility %in% "Eligible") %>% 
  select (c(id, name, first_name, college | contains("eligibility")))

write.csv(df_tee_inel, here::here("00_data", "processed_data", "eoi_mentee_ineligible.csv"), row.names = FALSE)

# save a subset with details about ineligible folks and those with outstanding questions
df_tor_inel <- df_tor %>% 
  filter(!eligibility %in% "Eligible") %>% 
  select (c(id, name, first_name, college | contains("eligibility")))

write.csv(df_tor_inel, here::here("00_data", "processed_data", "eoi_mentor_ineligible.csv"), row.names = FALSE)


```


```{r select eligible folks and join the tibbles step 1 of 2, include = FALSE}

# reassign these objects so we can easily see the original number of apps

df_tor_all <- df_tor
df_tee_all <-df_tee 

# keep mentors who are eligible

df_tor <- df_tor %>% 
  filter(eligibility %in% c("Eligible", "Information needed"))

# keep mentees who are eligible

df_tee <- df_tee %>% 
  filter(eligibility %in% c("Eligible", "Information needed"))

# change the type of variables to characters, so I can join them

df_tee <- df_tee %>% 
  mutate_if(is.numeric, as.character)

df_tor <- df_tor %>% 
  mutate_if(is.numeric, as.character)

#combine the eligible mentor and mentee tibbles into one

df <- full_join(df_tor, df_tee)

# then use this for additional processing


```


```{r make areas long, include = FALSE}

# RESHAPE THE DATA NOW, SO WE CAN INCLUDE THE NUMBERS OF FOLKS WHO WERE 
  # REMOVED BECAUSE THEY DIDN'T INCLUDE INFO ABOUT AREAS

# make the dataset with the ranks long

long <- df %>% 
  select (-c(first_name:title, work_phone:contract_end_date, why_mentor:mentoring_approach)) %>% 
  pivot_longer(
    cols = planning:other_2,
    names_to = "areas", 
    values_to = "rank"  
               )

# recode the ranks to yes or NA for interest 
  # [except some people entered other information, so I need to screen those out, as below]

# long2 <- long %>% 
#   mutate(interest = ifelse(rank %in% NA, NA, "yes")) 

long2 <- long %>%
  mutate(interest = case_when(rank == "no" ~ "no", 
                                 rank == "No" ~ "no",
                                 rank == "unsure" ~ "no",
                                 rank == "Maybe" ~ "no",
                                 rank == "Depending on discipline" ~ "no",
                                 rank == "Some experience" ~ "no",
                                 is.na(rank) ~ "no",
                                 TRUE ~ as.character("yes"))) %>% 
  mutate(interest = as.factor(interest))

```

```{r areas for all, include = FALSE}

# select relevant data
areas <- long2 %>% 
  select (c(id,`name and college`, name, college, level_recode, areas, rank, interest, role)) %>% 
  filter(interest == "yes")

# make this wide again and combine all of the interests into one cell. 
areas_wide <- areas %>% 
  select (-c(`name and college`, rank)) %>% 
  pivot_wider(names_from = "areas", values_from = areas)

# combine the areas into one column [if no one entered data for the 
  # `other_1` & `other_2` variables, then they need to be removed from from this 
  # process]

areas_wide <- areas_wide %>% 
  unite ("areas", planning:other_1, sep = ", ", remove = FALSE, na.rm = TRUE)

mentors_areas_wide <- areas_wide %>% 
      filter(role == "mentor")

mentees_areas_wide <- areas_wide %>% 
      filter(role == "mentee")

# we lose some folks who did not indicate any areas of interest; look for these by
  # using anti_join on the original datasets and the long-to-wide datasets; 
  # look at their original EOIs to ensure we aren't missing anything. 

no_areas_tor <- anti_join(df_tor, mentors_areas_wide, by = "id")
no_areas_tee <- anti_join(df_tee, mentees_areas_wide, by = "id")

# we lose one mentor because they did not provide any areas of interest
# we lose two mentees because they did not provide any areas of interest
# [see internal_notes for details]
# we then use these objects in the report to explain the total number of applications considered


# final number of mentees
total_tee <- length(unique(mentees_areas_wide$id))

# final number of mentors
total_tor <- length(unique(mentors_areas_wide$id))


# recode the df_tee and df_tor dataframes for completeness so they have only those who are valid
  # critical because we use these datasets below...
  # simplest way is to remove those that were not included in the mentees_area_wide from the df_tee

df_tee <- anti_join(df_tee, no_areas_tee) %>% 
    mutate_if(is.numeric, as.character) # confirm the columns are characters because they reverted to numbers

df_tor <- anti_join(df_tor, no_areas_tor) %>% 
    mutate_if(is.numeric, as.character)

```


```{r join the tibbles step 2 of 2, include = FALSE}

# combine the eligible mentor and mentee tibbles into one now that we have removed 
  # folks without areas listed

df <- full_join(df_tor, df_tee)


# when done assessing for eligibility and interests, write the data to new files
# row.names gets rid of the first column from the dataframe.

write.csv(df_tor, here::here("00_data", "processed_data", "eoi_mentor_eligible.csv"), row.names = FALSE)

write.csv(df_tee, here::here("00_data", "processed_data", "eoi_mentee_eligible.csv"), row.names = FALSE)

write.csv(df, here::here("00_data", "processed_data", "eoi_combined_eligible.csv"), row.names = FALSE)

# also write the areas wide data so we can load them in step 3

write.csv(mentors_areas_wide, here::here("00_data", "processed_data", "eoi_mentor_areas_wide.csv"), row.names = FALSE)

write.csv(mentees_areas_wide, here::here("00_data", "processed_data", "eoi_mentee_areas_wide.csv"), row.names = FALSE)


# these files will contain ONLY the eligible folks who provided info about their interests

```


```{r look at this info to find eligible folks to prioritise, include = FALSE, eval = FALSE}

# look at notes from the three previous years re: eligible but not selected folks who applied again this year

# look at eligible folks with contracts that end before 2025 (and thus can't participate next year).


df_tee_end_25 <- df_tee %>% 
  filter(str_detect(employment_length, "Fixed-term")) %>% 
  select(c(id, name, college, employment_length, contract_end_date, level_recode, employment_eligibility, level_eligibility, conferral_eligibility)) %>% 
  filter(grepl("25", contract_end_date))

# check anyone with contract end dates of 2026 too in case contracts end mid-year
df_tee_end_26 <- df_tee %>% 
  filter(str_detect(employment_length, "Fixed-term")) %>% 
  select(c(id, name, college, employment_length, contract_end_date, level_recode, employment_eligibility, level_eligibility, conferral_eligibility)) %>% 
  filter(grepl("26", contract_end_date))

rm(df_tee_end_25, df_tee_end_26)

# look at eligible folks with conferral dates between Feb 2017 and Feb 2018  (and thus can't participate next year)
  # double check the folks that had career interruptions.

df_phd <- df_tee %>% 
  select(c(id, name, college, conferral_month, conferral_year, interruptions, employment_eligibility, level_eligibility)) %>% 
  filter(as.numeric(conferral_year) <= 2018)

rm(df_phd)

```


We received `r original_n_tor` mentor applications and `r original_n_tee` mentee
applications. We screened this data for quality and then removed people who did not meet the
eligibility criteria for the program. Below are the screening steps:

  *  Duplicates: We had `r dup_tee` mentee applications with 
  multiple entries and `r dup_tor` mentor applications with multiple entries. 
  We removed incomplete copies of these duplicated applications.
  
  *  Withdrawals: We had `r withdraw_n_tee_pre` mentee applicant and `r withdraw_n_tor_pre` 
  mentor applicant withdraw from the program before matching took place. We then had
  `r withdraw_n_tee_pre` mentee applicant withdraw after matching took place. We will 
  keep their mentor as a potential back up in case another mentor withdraws after the program begins.
  
  *  Mentor eligibility:  We had `r level_no_n_tor` mentors who did not meet the 
  eligibility criteria for the program. These were likely mentees who accidentally 
  completed the incorrect form. 
  
  *  Mentee eligibility: We had
  `r sum(previous_no_n_tee, employ_no_n_tee, level_no_n_tee, contract_no_n_tee, phd_no_n_tee)` mentees 
  who did not meet the eligibility criteria for the program. Specifically, we 
  had `r previous_no_n_tee` application from a mentee who previously participated in the program,
  `r employ_no_n_tee` applications from people who did not hold academic positions, 
  `r phd_no_n_tee` application from people whose PhD was conferred more than 8 years ago or not 
  yet conferred, and
  `r contract_no_n_tee` application from someone whose contracts will end before November 2025. 

  *  Missing information: We removed `r nrow(no_areas_tee)`
  mentees from the data because they did not specify the areas for growth and
  `r nrow(no_areas_tor)` mentor because they did not identify the areas in which
  they could provide mentoring. 

As such, we completed the matching process on the remaining complete and eligible
applications from `r total_tor` mentors and `r total_tee` mentees. 


```{r college breakdown}

# how many eligible mentors per college

col_tor <- df_tor %>% 
  group_by(college) %>% 
  summarise(n = n()) %>% 
  mutate(perc = round(100*(n / sum(n)),2)) %>% 
  arrange(college) %>% 
  ungroup

# how many eligible mentees per college

col_tee <- df_tee %>% 
  group_by(college) %>% 
  summarise(n = n()) %>% 
  mutate(perc = round(100*(n / sum(n)),2)) %>% 
  arrange(college) %>% 
  ungroup

# keep the cell that is empty 
col <- full_join (col_tor, col_tee, by = "college") %>% 
  replace(is.na(.), 0)

# need to manually update the number of mentors and mentees in the header

col %>% 
knitr::kable (booktabs = T, linesep = "",
              col.names = c("College", "Number", "Percentage", "Number", "Percentage"), 
              caption = "Breakdown of Eligible Mentors and Mentees by College",
              align = 'lcc') %>% 
    add_header_above(c(" " = 1, "Mentors (n = 43)" = 2, "Mentees (n = 38)" = 2)) %>% 
  styler_pdf() 


```


\newpage

```{r ineligible mentee details to double check, include = FALSE}

# do not include this in the final report, but use it to double check our data

# identify the mentees in each college

df_tee_inel_details <- df_tee_inel %>% 
  select (c(college, name, id, eligibility)) %>% 
  ungroup

inel_tee <- nrow(df_tee_inel_details)

knitr::kable (
  df_tee_inel_details[, c(2,3,1,4)], # put the columns in the right order
  booktabs = T,
  linesep = "",
  col.names = c("Names", "Mentee Number", "College", "Eligibility"),
  caption =
    sprintf("Breakdown of Ineligible Mentees (n = %d)",
            inel_tee),
  align = 'lcc', 
  longtable = T #break the table over multiple pages
  ) %>%
  kable_styling(full_width = F, 
                latex_options = c("repeat_header", "hold_position", "striped"))
# use different styling to repeat the header

```


```{r ineligible mentor details to double check, include = FALSE}

# do not include this in the final report, but use it to double check our data

# identify the mentees in each college

df_tor_inel_details <- df_tor_inel %>% 
  select (c(college, name, id, eligibility)) %>% 
  ungroup

inel_tor <- nrow(df_tor_inel_details)

knitr::kable (
  df_tor_inel_details[, c(2,3,1,4)], # select only the names and levels columns
  booktabs = T,
  linesep = "",
  col.names = c("Names", "Mentee Number", "College", "Eligibility"),
  caption =
    sprintf("Breakdown of Ineligible Mentors (n = %d)",
            inel_tor),
  align = 'lcc', 
  longtable = T #break the table over multiple pages
  ) %>%
  kable_styling(full_width = F, 
                latex_options = c("repeat_header", "hold_position", "striped"))
# use different styling to repeat the header

```


```{r mentor details for matching process, include = FALSE}

#this will include their mentoring id number in the table for ease. We don't want
  # to include that for the enduring report though. For the report, make this false
  # and include the next chunk.

# identify the mentors in each college

# for now we will include eligibility in the table. We will remove this once eligibility
  # is confirmed

col_tor_details <- df_tor %>% 
  select (c(college, name, level_recode, id)) %>% 
  arrange(college, name) %>% 
  ungroup


knitr::kable (
  col_tor_details[, c(2,3,4)],
  booktabs = T,
  linesep = "",
  col.names = c("Name", "Level", "Mentor Number"),
  caption =
    sprintf("Breakdown of Mentors by College (n = %d)",
            total_tor),
  align = 'llcc'
  ) %>%
  pack_rows (index = table(col_tor_details$college)) %>%
  styler_pdf()

# if we want to save this as a png
# as_image(file = here::here("images", "mentors.png"))
```

```{r mentor details for report, include = TRUE}

#use this code for the report; it takes out their id number

# identify the mentors in each college

col_tor_details <- df_tor %>% 
  select (c(college, name, level_recode)) %>% 
  arrange(college) %>% 
  ungroup


knitr::kable (
  col_tor_details[, c(2,3)],
  booktabs = T,
  linesep = "",
  col.names = c("Name", "Level"),
  caption =
    sprintf("Breakdown of Mentors by College (n = %d)",
            total_tor),
  align = 'lcc'
  ) %>%
  pack_rows (index = table(col_tor_details$college)) %>%
  styler_pdf() 

```



```{r mentee details for matching process, include = FALSE}
# same re the report

# identify the mentees in each college

col_tee_details <- df_tee %>% 
  select (c(college, name, level_recode, id)) %>% 
  arrange(college, name) %>% 
  ungroup


knitr::kable (
  col_tee_details[, c(2,3, 4)], # select the names, levels and id columns
  booktabs = T,
  linesep = "",
  col.names = c("Name", "Level", "Mentee Number"),
  caption =
    sprintf("Breakdown of Mentees by College (n = %d)",
            total_tee),
  align = 'llc', 
  longtable = T #break the table over multiple pages
  ) %>%
  pack_rows (index = table(col_tee_details$college)) %>%
  kable_styling(full_width = F, 
                latex_options = c("repeat_header", "hold_position", "striped")) 
# use different styling to repeat the header

```

```{r mentee details for report, include = TRUE}

# identify the mentees in each college

col_tee_details <- df_tee %>% 
  select (c(college, name, level_recode)) %>% 
  arrange(college) %>% 
  ungroup


knitr::kable (
  col_tee_details[, c(2,3)], # select only the names and levels columns
  booktabs = T,
  linesep = "",
  col.names = c("Names", "Levels"),
  caption =
    sprintf("Breakdown of Mentees by College (n = %d)",
            total_tee),
  align = 'lc', 
  longtable = T #break the table over multiple pages
  ) %>%
  pack_rows (index = table(col_tee_details$college)) %>%
  kable_styling(full_width = F, 
                latex_options = c("repeat_header", "hold_position", "striped")) 
# use different styling to repeat the header

```
